requirements.txt
peewee~=3.14.0
numpy~=1.18.5
tensorflow~=2.3.0
opencv-python~=4.5.1.48
dlib~=19.21.1
gdown~=3.12.2
mtcnn~=0.1.0
pandas~=1.2.0
pillow~=8.1.0
pymysql~=1.0.2
requests~=2.25.1
send2trash~=1.5.0
project-pkgs~=0.0.5
tensorflow-gpu~=2.3.0
schedule~=0.6.0
enlighten~=1.7.2